Update paging initialization

+++ tile/kernel/page.c
@@ -51,7 +51,7 @@ void pte_insert(uint32_t* pte, uint32_t v_addr, uint32_t p_addr) {
 void init_paging() {
   int i;
   struct memory_map_block* b;
-  uint64_t b_end;
+  uint32_t b_end;
 
   for (i = 0; i < memory_map.memory->size; ++i) {
     b = &memory_map.memory->blocks[i];
